Add files via upload
Original research paper (authorship info in video) and video for the STAT 596 Individual Research Project (IRP). This is based on a paper entitled "Factor Copula Approaches for Assessing Spatially Dependent High Dimensional Risks." This project demonstrates the ability to select important information from a research article in the spatiotemporal analysis literature, and to distill it into a short talk. It also showcases presentation skills and LaTeX/Beamer competency.